Revert "Updating header h1 to use --swatch-headerh1-color var"
This reverts commit fe1836a4350eda5f53b7fd0dac49ad019cf2fbba.

After a discussion with Rounderhouse, we have decided to revert the
change to the header's shadow. The author of the commit has been
unresponsive and thus we have not been able to discuss the rationale
behind the change.

We are open to reimplementing once we can come to an agreement about
how to avoid issues this introduced in existing themes.