Extract language selector to class
Additional styling to the interwiki is added by BHL in the case that the
interwiki is not able to style itself, as is the case with the RU
interwiki which is used by most sites in the SCP network. This is not
the case for CN, which already has its own stylable interwiki and does
not need these adjustments. With the new interwiki made by EN (see
scpwiki/interwiki), the number of sites that need these adjustments is
expected to decrease; therefore it is not sustainable to use the site
language to specify which sites needs it. This commits extracts the
selector to a class, which scpwiki/interwiki encourages its users to
add; however this class is not currently present on CN so both sets of
selectors temporarily need to be used.