Move stylable selector from iframe to wrapper
The stylable class is only present on the wrapper, not the iframe
itself.

Just moving the :not(.interwiki-stylable) to its own selector will not
work, as this will select all elements that do not have that class even
if that element chain does contain such a class. Therefore
.scpnet-interwiki-wrapper needs to be present in the chain, and I moved
the CN selector there as well for consistency even though its location
doesn't matter.