Adding fonts.css to the import
In order to speed up loading of the theme css file, the fonts.css file which includes all the `@font-face` rules is set to also be imported into the main CSS files.

In order to do so, the relative src links needed to be replaced via PostCSS and a global `sed` command on the css.mk file.