Forum post editing title styling
Due to the fact that Wikidot inexplicably has a different structure for the Forum Post edit box as compared to the Article edit box, there needed to be specific styling for the Forum Post title box.

Unfortunately, since the "Post title:" text is floating on its own, not contained within the same box as the input box, it is currently not possible to make it so the "Post title:" text is responsive to whether the input box is currently active.