Fixing up ul/ol and adding dl styling
The ul/ol styling has been wonky for a while so this should fix it up. Also added dl (definition list) styling that fits in with the rest of BHL's visuals. Not sure how often Definition Lists are used on site but this should add a bit of visual flair to them.