Fix Font URL + Makefile Changes
Added a visitor to the LightningCSS build to ensure the Font URLs in `dist/css/black-highlighter.css` were correct.

Also made some changes to the Makefiles. It checks if the images in `dist` already exist and, if they do, only runs the image optimization if the filesizes are different.

I also added a `touch` to the CSS build since it seemed to be always running those `bun postcss` command but not updating the update time for the output file, for some reason. There's probably a better way to remedy this but it works for now.